Rename `ui.uicomponents` package to `ui.components`
Remove redundancy in name
