+++ ui/src/main/scala/org/narrativeandplay/hypedyn/ui/components/CentrePane.scala
@@ -1,4 +1,4 @@
-package org.narrativeandplay.hypedyn.ui.uicomponents
+package org.narrativeandplay.hypedyn.ui.components
 
 import scalafx.Includes._
 import scalafx.geometry.Orientation
