+++ ui/src/main/scala/org/narrativeandplay/hypedyn/ui/components/Toolbar.scala
@@ -1,4 +1,4 @@
-package org.narrativeandplay.hypedyn.ui.uicomponents
+package org.narrativeandplay.hypedyn.ui.components
 
 import java.lang
